refactor: split streaming orchestration helpers (#46)
